move streams module to data under metadata section
