+++ geoglows/__init__.py
@@ -2,15 +2,15 @@ import geoglows.bias
 import geoglows._plots as plots
 import geoglows.data
 import geoglows.analyze
-import geoglows.streams
 import geoglows.tables
 import geoglows.streamflow
 
 from ._constants import set_uri
+from ._constants import get_uri
 
 __all__ = [
-    'bias', 'plots', 'data', 'analyze', 'streams', 'tables', 'streamflow',
-    'set_uri'
+    'bias', 'plots', 'data', 'analyze', 'tables', 'streamflow',
+    'set_uri', 'get_uri'
 ]
 __version__ = '2.0.0'
 __author__ = 'Riley Hales'
